Arreglo zoom

+++ src/codigo/TP taller/TP taller/Vista/Dibujables/Dibujable.cpp
@@ -9,7 +9,7 @@ Dibujable::~Dibujable()
 void Dibujable::setColor(int* rgb, int a){
 }
 
-void Dibujable::dibujar(SDL_Renderer* renderer, int corrimiento, int escalaZoom, int posZoomX, int posZoomY){
+void Dibujable::dibujar(SDL_Renderer* renderer, int corrimiento, float escalaZoom, int posZoomX, int posZoomY){
 }
 
 DibujableSerializado Dibujable::getDibujableSerializado(int& tamano){
